compiler: fix memory-safety defects from the audit
Batched fixes for the memory-safety audit (14 high / 11 medium / low
sweep). All batches green on smoke; standard tier 568/568.

High - compiler itself:
- genmeta: gm_type_str accumulated snprintf would-be lengths, sending
  the write cursor past its stack buffer on truncated type names
- irgen_call: string.Format placeholder index wrapped negative and read
  args.items out of bounds; huge digit runs now stay literal text
- irgen_builtins: named-argument rebinding wrote NULL into the AST arg
  list when a named arg landed past a defaulted parameter (zanc crash /
  silent params miscompile); reject the shape with a diagnostic
- irgen_expr: query skip_stack[16] pushed without bound (16+ plain joins)

High - generated code:
- irgen_stmt: non-constant case labels reached LLVMConstIntGetSExtValue;
  diagnosed and routed to the switch end instead
- irgen_arc: array release cache keyed on an ambiguous, truncatable
  mangled token (string[][] vs class stringA); mix in a structure hash
- irgen_async via irgen_stmt: a same-named local of a different type
  reused the scanned frame slot and type-confused both variables
- checker: new int[n] {...} element count vs dimension product is now a
  compile error (the lowering writes elements with no runtime bound)
- irgen_builtins: string.Join strlen'd null list elements
- irgen_reflect: a 5-parameter static method thunk read one slot past
  the caller's fixed argument array
- irgen_call: File.Copy freed an RC-header pointer with libc free,
  fopen results were unchecked, negative ftell drove the copy

Medium:
- incremental: a truncated manifest stored half-initialized units with
  garbage dep pointers; loads now discard from the first bad record and
  saves go through tmp+rename
- binder/nsresolve: tuple-signature and namespace-join buffers silently
  truncated into identity collisions; full-length allocation or hash
  disambiguation
- irgen_expr: ToStr(off,len) clamped into the buffer, NativeMemory
  GetString/PutString guard negative lengths, ToBytes normalizes null
- irgen_stmt: pure type-pattern cases are no longer mistaken for the
  default arm of a pattern switch
- parser: accessor-name synthesis strdup'd past its 4-byte source;
  parse_block gained a statement depth cap
- lexer: lexer_advance guards EOF, preprocessor !/( evaluation got a
  recursion cap
- irgen_abi: sizes computed in 64-bit with saturation (32-bit long wrap
  misclassified MEMORY aggregates)
- genrun: double fclose on write-failure paths; embedres rejects >=4GiB
  resources instead of truncating LLVM array lengths
- irgen_call: Math.Abs/Max diagnose pointer operands instead of calling
  GetIntTypeWidth on them

Low/defensive:
- const folding and INT64_MIN negation done in unsigned arithmetic
- checker member loops null-guard entries like the rest of the file
- winres .ico bounds compared in 64 bits; version compare overflow-safe
  and sign-free; stdlib_ext JSON reader has a nesting cap

Tests: diag_named_args_gap, diag_array_init_count_mismatch,
conformance_regress_format_join_bounds (+determinism/leakcheck twins).

+++ src/compiler/winres.c
@@ -118,7 +118,7 @@ static int icon_object(const unsigned char *ico, size_t ico_len,
         ents[i].bits = (unsigned short)rd16(e + 6);
         ents[i].bytes = rd32(e + 8);
         ents[i].offset = rd32(e + 12);
-        if ((size_t)ents[i].offset + ents[i].bytes > ico_len) {
+        if ((unsigned long long)ents[i].offset + ents[i].bytes > ico_len) {
             free(ents); return 2;
         }
     }
